+++ NEWS.md
@@ -9,6 +9,8 @@
 * Added `suffix_format` argument to `fetch_data()` to specify the format of suffix in the query file. This is useful when you want to use a different format of suffix in the query file.
 * Enhanced the documentation of `fetch_data()`.
 * Let package not depend on dplyr, tidyr and purrr packages (#84).
+* Exported more targets factory functions: `tar_prep_proj()`, `tar_fetch_data()`, `tar_prep_raw()`.
+* Do not add `progress_hash` objects when `check_progress` is set to `FALSE` in `tar_prep_iquizoo()`.
 
 # tarflow.iquizoo 3.9.3